convert back to Node ESM to align with standards
If there is an issue, we will make it work. New tools already support it
(Webpack 5 supports 100% of Node ESM module resolution for example). We
will be sticking to a subset of Node ESM that is compatible with Browser
ESM (everything except bare module specifiers will be compatible, but
all standard build tools know how to handle them anyways). For non-build
systems, we will generate import maps.